# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API AppRestaurant",
        "description": "Listado de URI´S de la API restaurante",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/restaurants": {
            "get": {
                "tags": [
                    "Restaurant"
                ],
                "summary": "Listado de restaurants",
                "operationId": "0ef65917394f8eb77311d1d085ff31e4",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Aderson"
                                                    },
                                                    "imagen_path": {
                                                        "type": "string",
                                                        "example": "1.png"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Datos"
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "example": "etc"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "60020601"
                                                    },
                                                    "categori_id": {
                                                        "type": "sinteger",
                                                        "example": "1"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Restaurant"
                ],
                "summary": "Registrar la información del restautant",
                "operationId": "938db3c3288b69d111a27774178532ab",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "imagen_path": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "address": {
                                                "type": "string",
                                                "example": "etc"
                                            },
                                            "phone": {
                                                "type": "string",
                                                "example": "60020601"
                                            },
                                            "categori_id": {
                                                "type": "integer",
                                                "example": "1"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Aderson Felix",
                                    "imagen_path": "1.png",
                                    "description": "datos",
                                    "address": "datos",
                                    "phone": "60020601",
                                    "category_id": "1"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Aderson Felix"
                                        },
                                        "imagen_path": {
                                            "type": "string",
                                            "example": "1.png"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "pasta"
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": "datos"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "60020601"
                                        },
                                        "category_id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE CONTENT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The description field is required."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Objeto de errores"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}